認可機能実装

+++ 5.7/app/Http/Controllers/PostController.php
@@ -64,6 +64,7 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
+        $this->authorize('edit', $post);
         return view('posts.edit', ['post' => $post]);
     }
 
@@ -76,6 +77,7 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+        $this->authorize('edit', $post);
         $post->title = $request->title;
         $post->body = $request->body;
         $post->save();
@@ -90,6 +92,7 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        $this->authorize('edit', $post);
         $post->delete();
         return redirect('posts');
     }
